--- src/main/java/com/mulesoft/connectors/internal/operations/obsolete_InferenceOperations.java
@@ -1,172 +1,0 @@
-package com.mulesoft.connectors.internal.operations;
-
-import com.mulesoft.connectors.internal.api.metadata.LLMResponseAttributes;
-import com.mulesoft.connectors.internal.config.obsolete_InferenceConfiguration;
-import com.mulesoft.connectors.internal.exception.InferenceErrorType;
-import com.mulesoft.connectors.internal.utils.obsolete_ConnectionUtils;
-import com.mulesoft.connectors.internal.utils.obsolete_PayloadUtils;
-import com.mulesoft.connectors.internal.utils.obsolete_ResponseUtils;
-import org.json.JSONArray;
-import org.json.JSONObject;
-import org.mule.runtime.extension.api.annotation.Alias;
-import org.mule.runtime.extension.api.annotation.metadata.fixed.OutputJsonType;
-import org.mule.runtime.extension.api.annotation.param.Config;
-import org.mule.runtime.extension.api.annotation.param.Content;
-import org.mule.runtime.extension.api.annotation.param.MediaType;
-import org.mule.runtime.extension.api.annotation.param.display.DisplayName;
-import org.mule.runtime.extension.api.annotation.param.display.Summary;
-import org.mule.runtime.extension.api.exception.ModuleException;
-import org.mule.runtime.extension.api.runtime.operation.Result;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.*;
-import java.net.URL;
-
-import static org.mule.runtime.extension.api.annotation.param.MediaType.APPLICATION_JSON;
-
-/**
- * This class contains operations for the inference connector.
- * Each public method represents an extension operation.
- */
-public class obsolete_InferenceOperations {
-    private static final Logger LOGGER = LoggerFactory.getLogger(obsolete_InferenceOperations.class);
-    private static final String ERROR_MSG_FORMAT = "%s result error";
-
-    /**
-     * Chat completions by messages array including system, users messages i.e. conversation history
-     * @param configuration the connector configuration
-     * @param messages the conversation history as a JSON array
-     * @return result containing the LLM response
-     * @throws ModuleException if an error occurs during the operation
-     */
-    @MediaType(value = APPLICATION_JSON, strict = false)
-    @Alias("Chat-completions")
-    @DisplayName("[Chat] Completions")
-    @OutputJsonType(schema = "api/response/Response.json")
-    @Summary("Native chat completion operation")
-    public Result<InputStream, LLMResponseAttributes> chatCompletion(
-            @Config obsolete_InferenceConfiguration configuration,
-            @Content InputStream messages) throws ModuleException {
-        try {
-            JSONArray messagesArray = obsolete_PayloadUtils.parseInputStreamToJsonArray(messages);
-            URL chatCompUrl = obsolete_ConnectionUtils.getConnectionURLChatCompletion(configuration);
-
-            LOGGER.debug("Chatting with {}", chatCompUrl);
-
-            JSONObject payload = obsolete_PayloadUtils.buildPayload(configuration, messagesArray, null);
-            String response = obsolete_ConnectionUtils.executeREST(chatCompUrl, configuration, payload.toString());
-
-            LOGGER.debug("Chat completions result {}", response);
-            return obsolete_ResponseUtils.processLLMResponse(response, configuration);
-        } catch (Exception e) {
-            LOGGER.error("Error in chat completions: {}", e.getMessage(), e);
-            throw new ModuleException(String.format(ERROR_MSG_FORMAT, "Chat completions"),
-                    InferenceErrorType.CHAT_COMPLETION, e);
-        }
-    }
-
-    /**
-     * Simple chat answer for a single prompt
-     * @param configuration the connector configuration
-     * @param prompt the user's prompt
-     * @return result containing the LLM response
-     * @throws ModuleException if an error occurs during the operation
-     */
-    @MediaType(value = APPLICATION_JSON, strict = false)
-    @Alias("Chat-answer-prompt")
-    @DisplayName("[Chat] Answer Prompt")
-    @OutputJsonType(schema = "api/response/Response.json")
-    @Summary("Simple chat answer prompt")
-    public Result<InputStream, LLMResponseAttributes> chatAnswerPrompt(
-            @Config obsolete_InferenceConfiguration configuration,
-            @Content String prompt) throws ModuleException {
-        try {        
-        	JSONObject payload = obsolete_PayloadUtils.buildChatAnswerPromptPayload(configuration, prompt);
-
-            URL chatCompUrl = obsolete_ConnectionUtils.getConnectionURLChatCompletion(configuration);
-            String response = obsolete_ConnectionUtils.executeREST(chatCompUrl, configuration, payload.toString());
-
-            LOGGER.debug("Chat answer prompt result {}", response);
-            return obsolete_ResponseUtils.processLLMResponse(response, configuration);
-        } catch (Exception e) {
-            LOGGER.error("Error in chat answer prompt: {}", e.getMessage(), e);
-            throw new ModuleException(String.format(ERROR_MSG_FORMAT, "Chat answer prompt"),
-                    InferenceErrorType.CHAT_COMPLETION, e);
-        }
-    }
-
-    /**
-     * Define a prompt template with instructions and data
-     * @param configuration the connector configuration
-     * @param template the template string
-     * @param instructions instructions for the LLM
-     * @param data the primary data content
-     * @return result containing the LLM response
-     * @throws ModuleException if an error occurs during the operation
-     */
-    @MediaType(value = APPLICATION_JSON, strict = false)
-    @Alias("Agent-define-prompt-template")
-    @DisplayName("[Agent] Define Prompt Template")
-    @OutputJsonType(schema = "api/response/Response.json")
-    @Summary("Define a prompt template with instructions, and data ")
-    public Result<InputStream, LLMResponseAttributes> promptTemplate(
-            @Config obsolete_InferenceConfiguration configuration,
-            @Content String template,
-            @Content String instructions,
-            @Content(primary = true) String data) throws ModuleException {
-        try {
-        	        	
-        	JSONObject payload = obsolete_PayloadUtils.buildPromptTemplatePayload(configuration, template, instructions, data);
-
-            URL chatCompUrl = obsolete_ConnectionUtils.getConnectionURLChatCompletion(configuration);
-            String response = obsolete_ConnectionUtils.executeREST(chatCompUrl, configuration, payload.toString());
-
-            LOGGER.debug("Agent define prompt template result {}", response);
-            return obsolete_ResponseUtils.processLLMResponse(response, configuration);
-        } catch (Exception e) {
-            LOGGER.error("Error in agent define prompt template: {}", e.getMessage(), e);
-            throw new ModuleException(String.format(ERROR_MSG_FORMAT, "Agent define prompt template"),
-                    InferenceErrorType.CHAT_COMPLETION, e);
-        }
-    }
-
-    /**
-     * Define a tools template with instructions, data and tools
-     * @param configuration the connector configuration
-     * @param template the template string
-     * @param instructions instructions for the LLM
-     * @param data the primary data content
-     * @param tools tools configuration as a JSON array
-     * @return result containing the LLM response
-     * @throws ModuleException if an error occurs during the operation
-     */
-    @MediaType(value = APPLICATION_JSON, strict = false)
-    @Alias("Tools-native-template")
-    @DisplayName("[Tools] Native Template (Reasoning only)")
-    @OutputJsonType(schema = "api/response/Response.json")
-    @Summary("Define a prompt template with instructions, data and tools")
-    public Result<InputStream, LLMResponseAttributes> toolsTemplate(
-            @Config obsolete_InferenceConfiguration configuration,
-            @Content String template,
-            @Content String instructions,
-            @Content(primary = true) String data,
-            @Content @Summary("JSON Array defining the tools set to be used in the template so that the LLM can use them if required") InputStream tools) throws ModuleException {
-        
-    	try {
-        	        	
-        	JSONObject payload = obsolete_PayloadUtils.buildToolsTemplatePayload(configuration, template, instructions, data, tools);
-        	
-            URL chatCompUrl = obsolete_ConnectionUtils.getConnectionURLChatCompletion(configuration);
-            String response = obsolete_ConnectionUtils.executeREST(chatCompUrl, configuration, payload.toString());
-
-            LOGGER.debug("Tools use native template result {}", response);
-            return obsolete_ResponseUtils.processToolsResponse(response, configuration);
-        } catch (Exception e) {
-            LOGGER.error("Error in tools use native template: {}", e.getMessage(), e);
-            throw new ModuleException(String.format(ERROR_MSG_FORMAT, "Tools use native template"),
-                    InferenceErrorType.CHAT_COMPLETION, e);
-        }
-    }
-
-}
