--- src/main/java/com/mulesoft/connectors/internal/operations/obsolete_VisionOperations.java
@@ -1,77 +1,0 @@
-package com.mulesoft.connectors.internal.operations;
-
-import com.mulesoft.connectors.internal.api.metadata.LLMResponseAttributes;
-import com.mulesoft.connectors.internal.config.obsolete_InferenceConfiguration;
-import com.mulesoft.connectors.internal.config.obsolete_VisionConfiguration;
-import com.mulesoft.connectors.internal.exception.InferenceErrorType;
-import com.mulesoft.connectors.internal.utils.obsolete_ConnectionUtils;
-import com.mulesoft.connectors.internal.utils.obsolete_PayloadUtils;
-import com.mulesoft.connectors.internal.utils.obsolete_ProviderUtils;
-import com.mulesoft.connectors.internal.utils.obsolete_ResponseUtils;
-import org.json.JSONArray;
-import org.json.JSONObject;
-import org.mule.runtime.extension.api.annotation.Alias;
-import org.mule.runtime.extension.api.annotation.metadata.fixed.OutputJsonType;
-import org.mule.runtime.extension.api.annotation.param.Config;
-import org.mule.runtime.extension.api.annotation.param.Content;
-import org.mule.runtime.extension.api.annotation.param.MediaType;
-import org.mule.runtime.extension.api.annotation.param.display.DisplayName;
-import org.mule.runtime.extension.api.annotation.param.display.Summary;
-import org.mule.runtime.extension.api.exception.ModuleException;
-import org.mule.runtime.extension.api.runtime.operation.Result;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.InputStream;
-import java.net.URL;
-
-
-import static com.mulesoft.connectors.internal.utils.obsolete_PayloadUtils.createRequestImageURL;
-import static org.mule.runtime.extension.api.annotation.param.MediaType.APPLICATION_JSON;
-
-/**
- * This class contains operations for the inference connector.
- * Each public method represents an extension operation.
- */
-public class obsolete_VisionOperations {
-    private static final Logger LOGGER = LoggerFactory.getLogger(obsolete_VisionOperations.class);
-    private static final String ERROR_MSG_FORMAT = "%s result error";
-
-    /**
-     * Chat completions by messages array including system, users messages i.e. conversation history
-     * @param configuration the connector configuration
-     * @param prompt the users prompt
-     * @param imageUrl the image Url to be sent to the Vision Model
-     * @return result containing the LLM response
-     * @throws ModuleException if an error occurs during the operation
-     */
-    @MediaType(value = APPLICATION_JSON, strict = false)
-    @Alias("Read-image")
-    @DisplayName("[Image] Read by (Url or Base64)")
-    @OutputJsonType(schema = "api/response/Response.json")
-    public Result<InputStream, LLMResponseAttributes> readImage(
-            @Config obsolete_VisionConfiguration configuration,
-            @Content String prompt,
-            @Content(primary = true) @DisplayName("Image") @Summary("An Image URL or a Base64 Image") String imageUrl) throws ModuleException {
-        try {
-
-            JSONArray messagesArray = createRequestImageURL(configuration.getInferenceType(), prompt, imageUrl);
-
-            obsolete_InferenceConfiguration inferenceConfig = obsolete_ProviderUtils.convertToInferenceConfig(configuration);
-
-            URL chatCompUrl = obsolete_ConnectionUtils.getConnectionURLChatCompletion(inferenceConfig);
-            LOGGER.debug("Chatting with {}", chatCompUrl);
-
-            JSONObject payload = obsolete_PayloadUtils.buildPayload(inferenceConfig, messagesArray, null);
-
-            String response = obsolete_ConnectionUtils.executeREST(chatCompUrl, inferenceConfig, payload.toString());
-
-            LOGGER.debug("Read Image result {}", response);
-            return obsolete_ResponseUtils.processLLMResponse(response, inferenceConfig);
-        } catch (Exception e) {
-            LOGGER.error("Error in Read Image: {}", e.getMessage(), e);
-            throw new ModuleException(String.format(ERROR_MSG_FORMAT, "Read Image"),
-                    InferenceErrorType.VISION, e);
-        }
-    }
-}
