--- src/main/java/com/mulesoft/connectors/internal/utils/obsolete_PayloadUtils.java
@@ -1,581 +1,0 @@
-package com.mulesoft.connectors.internal.utils;
-
-import com.mulesoft.connectors.internal.config.obsolete_InferenceConfiguration;
-import com.mulesoft.connectors.internal.constants.InferenceConstants;
-
-import java.io.*;
-import java.net.URLConnection;
-import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-import java.util.Base64;
-
-import org.json.JSONArray;
-import org.json.JSONObject;
-
-/**
- * Utility class for payload operations.
- */
-public class obsolete_PayloadUtils {
-    private static final String[] NO_TEMPERATURE_MODELS = {"o3-mini", "o1", "o1-mini"};
-
-    /**
-     * Build the payload for the API request
-     * @param configuration the connector configuration
-     * @param messagesArray the messages array
-     * @param toolsArray the tools array (can be null)
-     * @return the payload as a JSON object
-     */
-    public static JSONObject buildPayload(obsolete_InferenceConfiguration configuration, JSONArray messagesArray, JSONArray toolsArray) {
-        JSONObject payload = new JSONObject();
-        
-		if ("VERTEX_AI_EXPRESS".equalsIgnoreCase(configuration.getInferenceType())) {
-			//add contents to the payload
-	        payload.put(InferenceConstants.CONTENTS, messagesArray);
-	        
-	        //create the generationConfig
-	        JSONObject generationConfig = buildVertexAIGenerationConfig(configuration);	        
-	        
-	        //add generationConfig to the payload
-	        payload.put(InferenceConstants.GENERATION_CONFIG, generationConfig);	        
-
-		} else {
-		
-	        if (!"AZURE_OPENAI".equals(configuration.getInferenceType())) {
-	            payload.put(InferenceConstants.MODEL, configuration.getModelName());
-	        }
-	        payload.put(InferenceConstants.MESSAGES, messagesArray);
-	
-	        // Different max token parameter names for different providers
-	        if ("GROQ".equalsIgnoreCase(configuration.getInferenceType()) ||
-	                "OPENAI".equalsIgnoreCase(configuration.getInferenceType())) {
-	            payload.put(InferenceConstants.MAX_COMPLETION_TOKENS, configuration.getMaxTokens());
-	        } else {
-	            payload.put(InferenceConstants.MAX_TOKENS, configuration.getMaxTokens());
-	        }
-	
-	        // Some models don't support temperature/top_p parameters
-	        String modelName = configuration.getModelName();
-	        if (!Arrays.asList(NO_TEMPERATURE_MODELS).contains(modelName)) {
-	            payload.put(InferenceConstants.TEMPERATURE, configuration.getTemperature());
-	            payload.put(InferenceConstants.TOP_P, configuration.getTopP());
-	        }
-	
-	        // Add tools array if provided
-	        if (toolsArray != null && !toolsArray.isEmpty()) {
-	            payload.put(InferenceConstants.TOOLS, toolsArray);
-	        }
-	
-	        // Special handling for Ollama's and Azure OpenAI's stream parameter
-	        if ("OLLAMA".equals(configuration.getInferenceType()) || "AZURE_OPENAI".equals(configuration.getInferenceType())) {
-	            payload.put("stream", false);
-	        }
-		}
-		
-        return payload;
-    }
-
-    /**
-     * Build the payload for the API request
-     * @param configuration the connector configuration
-     * @param requestJson the payload with prompt
-     * @return the payload as a JSON object
-     */
-    public static JSONObject buildPayloadImageGeneration(obsolete_InferenceConfiguration configuration, JSONObject requestJson) {
-        JSONObject payload = requestJson;
-
-        if ("OPENAI".equalsIgnoreCase(configuration.getInferenceType())) {
-            //add contents to the payload
-            payload.put("model", configuration.getModelName());
-        }
-
-        return payload;
-    }
-    
-    /**
-     * Build the payload for the Vertex AI API request
-     * @param configuration the connector configuration
-     * @param prompt the prompt
-     * @return the payload as a JSON object
-     */
-    public static JSONObject buildVertexAIPayload(obsolete_InferenceConfiguration configuration, String prompt,
-                                                  JSONArray safetySettings, JSONObject systemInstruction, JSONArray tools) {
-        JSONObject payload = new JSONObject();
-        
-        //create the parts of the contents
-        JSONArray partsArray = new JSONArray();
-        JSONObject usersPrompt = new JSONObject();
-        usersPrompt.put("text", prompt);
-        partsArray.put(usersPrompt);
-        
-        // Create the user role object
-        JSONObject userContent = new JSONObject();
-        userContent.put("role", "user");
-        userContent.put("parts", partsArray);
-        
-        // Create the contents array and add the user content
-        JSONArray contentsArray = new JSONArray();
-        contentsArray.put(userContent);
-
-        //add contents to the payload
-        payload.put(InferenceConstants.CONTENTS, contentsArray);
-
-        //add system instruction if provided
-        if (systemInstruction != null) {
-            payload.put(InferenceConstants.SYSTEM_INSTRUCTION, systemInstruction);
-        }
-
-        //create the generationConfig
-        JSONObject generationConfig = buildVertexAIGenerationConfig(configuration);	        
-        
-        //add generationConfig to the payload
-        payload.put(InferenceConstants.GENERATION_CONFIG, generationConfig);	        
-        
-        //add safety settings if provided
-        if (safetySettings != null && !safetySettings.isEmpty()) {
-            payload.put(InferenceConstants.SAFETY_SETTINGS, safetySettings);
-        }
-        
-        //add tools if provided
-        if (tools != null && !tools.isEmpty()) {
-            payload.put(InferenceConstants.TOOLS, tools);
-        }
-
-        return payload;
-    }
-    
-    
-
-
-    /**
-     * Parse an input stream to a JSON array
-     * @param inputStream the input stream to parse
-     * @return the parsed JSON array
-     * @throws IOException if an error occurs during parsing
-     */
-    public static JSONArray parseInputStreamToJsonArray(InputStream inputStream) throws IOException {
-        if (inputStream == null) {
-            return new JSONArray();
-        }
-
-        try (InputStreamReader reader = new InputStreamReader(inputStream, StandardCharsets.UTF_8);
-             BufferedReader bufferedReader = new BufferedReader(reader)) {
-
-            StringBuilder stringBuilder = new StringBuilder();
-            String line;
-            while ((line = bufferedReader.readLine()) != null) {
-                stringBuilder.append(line);
-            }
-
-            String jsonString = stringBuilder.toString().trim();
-            if (jsonString.isEmpty()) {
-                return new JSONArray();
-            }
-
-            return new JSONArray(jsonString);
-        }
-    }
-
-    /**
-     * Creates a messages array with system prompt and user message
-     * @param configuration the connector configuration
-     * @param systemContent content for the system/assistant message
-     * @param userContent content for the user message
-     * @return JSONArray containing the messages
-     */
-    public static JSONArray createMessagesArrayWithSystemPrompt(
-            obsolete_InferenceConfiguration configuration, String systemContent, String userContent) {
-        JSONArray messagesArray = new JSONArray();
-
-        // Create system/assistant message based on provider
-        JSONObject systemMessage = new JSONObject();
-        systemMessage.put("role", obsolete_ProviderUtils.isAnthropic(configuration) ? "assistant" : "system");
-        systemMessage.put("content", systemContent);
-        messagesArray.put(systemMessage);
-
-        // Create user message
-        JSONObject userMessage = new JSONObject();
-        userMessage.put("role", "user");
-        userMessage.put("content", userContent);
-        messagesArray.put(userMessage);
-
-        return messagesArray;
-    }
-
-
-    public static JSONArray createRequestImageURL(String provider, String prompt, String imageUrl) throws IOException {
-
-        if (provider.equalsIgnoreCase("ANTHROPIC")) {
-            return createAnthropicImageURLRequest(prompt, imageUrl);
-        } else if (provider.equalsIgnoreCase("OLLAMA")) {
-            return createOllamaImageURLRequest(prompt, imageUrl);
-        } else if (provider.equalsIgnoreCase("VERTEX_AI_EXPRESS")) {
-            return createVertexAIExpressImageURLRequest(prompt, imageUrl);
-        } else {
-            return createImageURLRequest(prompt, imageUrl);
-        }
-    }
-
-    /**
-     * Creates a messages array with system prompt and user message
-     * @param prompt of the user
-     * @param imageUrl of the image
-     * @return JSONArray containing the messages
-     */
-    private static JSONArray createImageURLRequest(String prompt, String imageUrl) throws IOException {
-        JSONArray messagesArray = new JSONArray();
-        JSONObject userMessage = new JSONObject();
-        userMessage.put("role", "user");
-        JSONArray contentArray = new JSONArray();
-
-        JSONObject textContent = new JSONObject();
-        textContent.put("type", "text");
-        textContent.put("text", prompt);
-        contentArray.put(textContent);
-
-        JSONObject imageContent = new JSONObject();
-        imageContent.put("type", "image_url");
-        JSONObject imageMessage = new JSONObject();
-        if (isBase64String(imageUrl)) {
-            imageMessage.put("url", "data:" + getMimeType(imageUrl) + ";base64," + imageUrl);
-        } else{
-            imageMessage.put("url", imageUrl);
-        }
-        imageContent.put("image_url", imageMessage);
-        contentArray.put(imageContent);
-
-        userMessage.put("content", contentArray);
-        messagesArray.put(userMessage);
-
-
-        return messagesArray;
-    }
-
-    /**
-     * Creates a messages array with system prompt and user message
-     * @param prompt of the user
-     * @param imageUrl of the image
-     * @return JSONArray containing the messages
-     */
-    private static JSONArray createOllamaImageURLRequest(String prompt, String imageUrl) throws IOException {
-        // Create messages array
-        JSONArray messagesArray = new JSONArray();
-
-        // Create user message object
-        JSONObject userMessage = new JSONObject();
-        userMessage.put("role", "user");
-        userMessage.put("content", prompt);
-
-        // Create images array
-        JSONArray imagesArray = new JSONArray();
-
-        // Add the image (either base64 or URL)
-        if (isBase64String(imageUrl)) {
-            // For base64 images, no need to add data: prefix as it's handled directly
-            imagesArray.put(imageUrl);
-        } else {
-            // For URL images
-            imagesArray.put(imageUrl);
-        }
-
-        // Add images array to user message
-        userMessage.put("images", imagesArray);
-
-        // Add the message to the array
-        messagesArray.put(userMessage);
-
-        return messagesArray;
-    }
-
-
-    /**
-     * Creates a messages array with system prompt and user message
-     * @param prompt of the user
-     * @param imageUrl of the image
-     * @return JSONArray containing the messages
-     */
-    private static JSONArray createAnthropicImageURLRequest(String prompt, String imageUrl) throws IOException {
-        JSONArray messagesArray = new JSONArray();
-        JSONObject userMessage = new JSONObject();
-        userMessage.put("role", "user");
-
-        JSONArray contentArray = new JSONArray();
-
-        JSONObject imageContent = new JSONObject();
-        imageContent.put("type", "image");
-        JSONObject imageSource = new JSONObject();
-        if (isBase64String(imageUrl)) {
-            imageSource.put("type", "base64");
-            imageSource.put("media_type", getMimeType(imageUrl));
-            imageSource.put("data", imageUrl);
-        } else{
-            imageSource.put("type", "url");
-            imageSource.put("url", imageUrl);
-        }
-
-        imageContent.put("source", imageSource);
-        contentArray.put(imageContent);
-
-        JSONObject textContent = new JSONObject();
-        textContent.put("type", "text");
-        textContent.put("text", prompt);
-        contentArray.put(textContent);
-
-        userMessage.put("content", contentArray);
-        messagesArray.put(userMessage);
-
-        return messagesArray;
-    }
-    
-    /**
-     * Creates a messages array with system prompt and base64 contents or URL of the image for Vertex AI Express
-     * @param prompt of the user
-     * @param imageUrl of the image
-     * @return JSONArray containing the messages
-     */
-    private static JSONArray createVertexAIExpressImageURLRequest(String prompt, String imageUrl) throws IOException {
-
-    	JSONArray parts = new JSONArray();
-    	
-        if (isBase64String(imageUrl)) {
-            JSONObject inlineData = new JSONObject();
-            inlineData.put("mimeType", getMimeType(imageUrl));
-            inlineData.put("data", imageUrl);  // imageUrl contains base64 in this case
-
-            JSONObject inlineDataWrapper = new JSONObject();
-            inlineDataWrapper.put("inlineData", inlineData);
-
-            parts.put(inlineDataWrapper);
-        } else {
-            JSONObject fileData = new JSONObject();
-            fileData.put("mimeType", getMimeTypeFromUrl(imageUrl));
-            fileData.put("fileUri", imageUrl);  // imageUrl contains gs:// URI in this case
-
-            JSONObject fileDataWrapper = new JSONObject();
-            fileDataWrapper.put("fileData", fileData);
-
-            parts.put(fileDataWrapper);
-        }
-
-        // Add text part
-        JSONObject textPart = new JSONObject();
-        textPart.put("text", prompt);
-
-        parts.put(textPart);
-
-        // Add to contents array
-        JSONObject content = new JSONObject();
-        content.put("role", "user");
-        content.put("parts", parts);
-
-        JSONArray contents = new JSONArray();
-        contents.put(content);
-
-        return contents;
-    
-    }
-
-
-    
-    /**
-     * Build the payload for the chatAnswerPrompt request
-     * @param configuration The connector configuration
-     * @param prompt The prompt
-     * @return The payload as a JSON object
-     */
-    public static JSONObject buildChatAnswerPromptPayload(obsolete_InferenceConfiguration configuration, String prompt) {
-    	JSONObject payload;
-	
-		if ("VERTEX_AI_EXPRESS".equalsIgnoreCase(configuration.getInferenceType())) {
-	    	JSONArray safetySettings = new JSONArray(); // Empty array
-	    	JSONObject systemInstruction = new JSONObject(); // Empty object
-	    	JSONArray tools = new JSONArray(); // Empty array
-			payload = obsolete_PayloadUtils.buildVertexAIPayload(configuration, prompt, safetySettings, systemInstruction, tools);
-		} else {
-	        JSONArray messagesArray = new JSONArray();
-	        JSONObject usersPrompt = new JSONObject();
-	        usersPrompt.put("role", "user");
-	        usersPrompt.put("content", prompt);
-	        messagesArray.put(usersPrompt);
-	        payload = obsolete_PayloadUtils.buildPayload(configuration, messagesArray, null);
-		}
-	
-		return payload;
-    }
-    
-    /**
-     * Build the payload for the promptTemplate request
-     * @param configuration The connector configuration
-     * @param template The template string
-     * @param instructions The instructions string
-     * @param data The primary data content
-     * @return The payload as a JSON object
-     */
-    public static JSONObject buildPromptTemplatePayload(obsolete_InferenceConfiguration configuration, String template, String instructions, String data) {
-    	JSONObject payload;
-	
-	
-		if ("VERTEX_AI_EXPRESS".equalsIgnoreCase(configuration.getInferenceType())) {
-	    	//Create systemInstruction object
-	    	//Step 1: Wrap text in a part object
-	        JSONObject part = new JSONObject();
-	        part.put("text", template + " - " + instructions);
-	
-	        //Step 2: Create parts array
-	        JSONArray parts = new JSONArray();
-	        parts.put(part);
-	
-	        //Step 3: Create systemInstruction object
-	        JSONObject systemInstruction = new JSONObject();
-	        systemInstruction.put("parts", parts);
-	    	
-	    	JSONArray safetySettings = new JSONArray(); // Empty array
-
-	    	JSONArray tools = new JSONArray(); // Empty array
-
-	    	payload = obsolete_PayloadUtils.buildVertexAIPayload(configuration, data, safetySettings, systemInstruction, tools);
-		} else {
-	        JSONArray messagesArray = obsolete_PayloadUtils.createMessagesArrayWithSystemPrompt(
-	                configuration, template + " - " + instructions, data);
-	
-	        payload = obsolete_PayloadUtils.buildPayload(configuration, messagesArray, null);
-	
-		}
-
-		return payload;
-
-    }
-
-    /**
-     * Build the payload for the toolsTemplate request
-     * @param configuration The connector configuration
-     * @param template The template string
-     * @param instructions The instructions string
-     * @param data The primary data content
-     * @param tools The tools set to be used
-     * @return The payload as a JSON object
-     */
-    public static JSONObject buildToolsTemplatePayload(obsolete_InferenceConfiguration configuration, String template,
-                                                       String instructions, String data, InputStream tools) throws IOException {
-
-        	JSONObject payload;
-
-        	
-        	if ("VERTEX_AI_EXPRESS".equalsIgnoreCase(configuration.getInferenceType())) {
-            	
-            	//Create systemInstruction object
-            	//Step 1: Wrap text in a part object
-                JSONObject part = new JSONObject();
-                part.put("text", template + " - " + instructions);
-
-                //Step 2: Create parts array
-                JSONArray parts = new JSONArray();
-                parts.put(part);
-
-                //Step 3: Create systemInstruction object
-                JSONObject systemInstruction = new JSONObject();
-                systemInstruction.put("parts", parts);
-            	
-            	JSONArray toolsArray = obsolete_PayloadUtils.parseInputStreamToJsonArray(tools);
-
-            	JSONArray safetySettings = new JSONArray(); // Empty array
-
-            	payload = obsolete_PayloadUtils.buildVertexAIPayload(configuration, data, safetySettings, systemInstruction, toolsArray);
-        	} else {
-        	
-        	
-        		JSONArray toolsArray = obsolete_PayloadUtils.parseInputStreamToJsonArray(tools);
-        		JSONArray messagesArray = obsolete_PayloadUtils.createMessagesArrayWithSystemPrompt(
-        				configuration, template + " - " + instructions, data);
-        		payload = obsolete_PayloadUtils.buildPayload(configuration, messagesArray, toolsArray);
-            }
-    
-    		return payload;
-    }
-
-    public static boolean isBase64String(String str) {
-        if (str == null || str.length() % 4 != 0 || !str.matches("^[A-Za-z0-9+/]*={0,2}$")) {
-            return false;
-        }
-        try {
-            Base64.getDecoder().decode(str);
-            return true;
-        } catch (IllegalArgumentException e) {
-            return false;
-        }
-    }
-
-    public static String getMimeType(String base64String) throws IOException {
-        byte[] decodedBytes = Base64.getDecoder().decode(base64String);
-        ByteArrayInputStream inputStream = new ByteArrayInputStream(decodedBytes);
-        String mimeType = URLConnection.guessContentTypeFromStream(inputStream);
-        return mimeType != null ? mimeType : "image/jpeg";
-    }
-    
-    public static String getMimeTypeFromUrl(String imageUrl) {
-        if (imageUrl == null || imageUrl.trim().isEmpty()) {
-            return "image/jpeg"; // Default
-        }
-
-        imageUrl = imageUrl.toLowerCase().trim();
-
-        if (imageUrl.endsWith(".png")) {
-            return "image/png";
-        } else if (imageUrl.endsWith(".jpg") || imageUrl.endsWith(".jpeg")) {
-            return "image/jpeg";
-        }
-
-        // Add more types if needed
-        return "image/jpeg"; // Default fallback
-    }
-    
-    public static JSONObject buildVertexAIGenerationConfig(obsolete_InferenceConfiguration configuration) {
-	    //create the generationConfig
-	    JSONObject generationConfig = new JSONObject();
-	    generationConfig.put("responseModalities", new String[]{"TEXT"});
-	    generationConfig.put("temperature", configuration.getTemperature());
-	    generationConfig.put("maxOutputTokens", configuration.getMaxTokens());
-	    generationConfig.put("topP", configuration.getTopP());
-        
-	    return generationConfig;
-    }
-
-
-    public static JSONObject createRequestImageGeneration(String provider, String prompt) throws IOException {
-
-        if (provider.equalsIgnoreCase("OPENAI")) {
-            return createImageGenerationRequestOpenaI(prompt);
-        } else if(provider.equalsIgnoreCase("HUGGING_FACE")){
-            return createImageGenerationRequestHuggingface(prompt);
-        } else {
-            return createImageGenerationRequestOpenaI(prompt);
-        }
-    }
-
-    /**
-     * Creates a messages array with system prompt and user message
-     * @param prompt of the user
-     * @return JSONArray containing the messages
-     */
-    private static JSONObject createImageGenerationRequestOpenaI(String prompt) throws IOException {
-
-        JSONObject requestPayload = new JSONObject();
-        requestPayload.put("prompt", prompt);
-        requestPayload.put("response_format", "b64_json");
-
-        return requestPayload;
-    }
-
-    /**
-     * Creates a messages array with system prompt and user message
-     * @param prompt of the user
-     * @return JSONArray containing the messages
-     */
-    private static JSONObject createImageGenerationRequestHuggingface(String prompt) throws IOException {
-
-        JSONObject requestPayload = new JSONObject();
-        requestPayload.put("inputs", prompt);
-        return requestPayload;
-    }
-
-
-}
